requirements.txt
pyTelegramBotAPI==4.1.1
flask==2.0.2
python-decouple==3.5
gunicorn==20.1.0